# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Sanctuary API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudySessionStatus =
  (typeof StudySessionStatus)[keyof typeof StudySessionStatus];

export const StudySessionStatus = {
  active: "active",
  completed: "completed",
} as const;

export interface StudySession {
  id: number;
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  status: StudySessionStatus;
  packEarned?: boolean;
}

export interface CurrentSessionResponse {
  session: StudySession | null;
}

export interface Stats {
  totalStudySeconds: number;
  sessionsCompleted: number;
  cardsCollected: number;
  packsOpened: number;
  uniqueCards: number;
  legendaryCards: number;
  flashcardsCreated?: number;
  quizzesCreated?: number;
}

export type CardRarity = (typeof CardRarity)[keyof typeof CardRarity];

export const CardRarity = {
  common: "common",
  rare: "rare",
  holographic: "holographic",
  legendary: "legendary",
} as const;

export type CardType = (typeof CardType)[keyof typeof CardType];

export const CardType = {
  creature: "creature",
  quote: "quote",
  fact: "fact",
  achievement: "achievement",
  cosmetic: "cosmetic",
} as const;

export interface Card {
  id: number;
  name: string;
  rarity: CardRarity;
  type: CardType;
  description: string;
  imageUrl: string;
  element: string;
  /** @nullable */
  power?: number | null;
  /** @nullable */
  flavorText?: string | null;
}

export interface CollectedCard {
  id: number;
  cardId: number;
  card: Card;
  acquiredAt: string;
  /** @nullable */
  packOpeningId?: number | null;
}

export interface PackStatus {
  available: boolean;
  studySecondsRequired: number;
  currentStudySeconds: number;
  packsAvailable: number;
}

export interface PackOpenResult {
  packId: number;
  cards: CollectedCard[];
}

export type StudyMaterialType =
  (typeof StudyMaterialType)[keyof typeof StudyMaterialType];

export const StudyMaterialType = {
  notes: "notes",
  pdf: "pdf",
  slides: "slides",
  guide: "guide",
} as const;

export interface StudyMaterial {
  id: number;
  title: string;
  type: StudyMaterialType;
  content: string;
  createdAt: string;
  hasGeneratedContent: boolean;
  flashcardCount?: number;
  quizCount?: number;
}

export type CreateStudyMaterialBodyType =
  (typeof CreateStudyMaterialBodyType)[keyof typeof CreateStudyMaterialBodyType];

export const CreateStudyMaterialBodyType = {
  notes: "notes",
  pdf: "pdf",
  slides: "slides",
  guide: "guide",
} as const;

export interface CreateStudyMaterialBody {
  title: string;
  type: CreateStudyMaterialBodyType;
  content: string;
}

export interface GenerationResult {
  flashcardsGenerated: number;
  quizzesGenerated: number;
}

export interface Flashcard {
  id: number;
  materialId: number;
  front: string;
  back: string;
  /** @nullable */
  category: string | null;
  difficulty: number;
  /** @nullable */
  lastReviewedAt?: string | null;
  /** @nullable */
  nextReviewAt?: string | null;
  reviewCount?: number;
}

export interface ReviewFlashcardBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  difficulty: number;
}

export interface Quiz {
  id: number;
  materialId: number;
  question: string;
  options: string[];
  correctIndex: number;
  /** @nullable */
  explanation?: string | null;
  timesAnswered?: number;
  timesCorrect?: number;
}

export interface AnswerQuizBody {
  selectedIndex: number;
}

export interface AnswerResult {
  correct: boolean;
  correctIndex: number;
  /** @nullable */
  explanation: string | null;
}

export type ListFlashcardsParams = {
  materialId?: number;
};

export type ListQuizzesParams = {
  materialId?: number;
};